refactor: Restructure feature and state imports, consolidate geometry models under geojson-pydantic
