+++ libs/shared-types/python-src/debrief/types/features/__init__.py
@@ -1,14 +1,8 @@
-"""Debrief maritime feature types."""
+"""Feature types for maritime GeoJSON data."""
 
 from .track import DebriefTrackFeature
 from .point import DebriefPointFeature
 from .annotation import DebriefAnnotationFeature
-from .feature_collection import DebriefFeatureCollection, DebriefFeature
+from .debrief_feature_collection import DebriefFeatureCollection
 
-__all__ = [
-    "DebriefTrackFeature",
-    "DebriefPointFeature",
-    "DebriefAnnotationFeature",
-    "DebriefFeatureCollection",
-    "DebriefFeature",
-]
+__all__ = ['DebriefTrackFeature', 'DebriefPointFeature', 'DebriefAnnotationFeature', 'DebriefFeatureCollection']
