+++ libs/shared-types/python-src/debrief/types/features/track.py
@@ -1,54 +1,14 @@
 """Track Pydantic model for maritime track features."""
 
 from datetime import datetime
-from pydantic import BaseModel, Field, validator
+from pydantic import BaseModel, Field, field_validator
 from typing import List, Union, Literal, Optional, Any, Dict
 from enum import Enum
+from geojson_pydantic import Feature, LineString, MultiLineString
 
 
-class GeometryType(str, Enum):
-    """Geometry types for track features."""
-    LINESTRING = "LineString"
-    MULTILINESTRING = "MultiLineString"
-
-
-class LineStringGeometry(BaseModel):
-    """LineString geometry for track features."""
-    type: Literal["LineString"] = "LineString"
-    coordinates: List[List[float]] = Field(
-        ...,
-        min_length=2,
-        description="Array of coordinate positions"
-    )
-
-    @validator('coordinates')
-    def validate_coordinates(cls, v):
-        """Validate coordinate array structure."""
-        for coord in v:
-            if len(coord) < 2 or len(coord) > 3:
-                raise ValueError('Each coordinate must have 2 or 3 elements (lon, lat, [elevation])')
-        return v
-
-
-class MultiLineStringGeometry(BaseModel):
-    """MultiLineString geometry for track features."""
-    type: Literal["MultiLineString"] = "MultiLineString"
-    coordinates: List[List[List[float]]] = Field(
-        ...,
-        min_length=1,
-        description="Array of LineString coordinate arrays"
-    )
-
-    @validator('coordinates')
-    def validate_coordinates(cls, v):
-        """Validate coordinate array structure."""
-        for linestring in v:
-            if len(linestring) < 2:
-                raise ValueError('Each LineString must have at least 2 coordinate positions')
-            for coord in linestring:
-                if len(coord) < 2 or len(coord) > 3:
-                    raise ValueError('Each coordinate must have 2 or 3 elements (lon, lat, [elevation])')
-        return v
+# Geometry types are now provided by geojson-pydantic
+# LineString and MultiLineString are imported from geojson_pydantic
 
 
 class TrackProperties(BaseModel):
@@ -74,16 +34,13 @@ class TrackProperties(BaseModel):
         extra = "allow"  # Allow additional properties
 
 
-class DebriefTrackFeature(BaseModel):
+class DebriefTrackFeature(Feature[Union[LineString, MultiLineString], TrackProperties]):
     """A GeoJSON Feature representing a track with LineString or MultiLineString geometry."""
 
-    type: Literal["Feature"] = "Feature"
-    id: Union[str, int] = Field(
-        ...,
-        description="Unique identifier for this feature"
-    )
-    geometry: Union[LineStringGeometry, MultiLineStringGeometry]
-    properties: TrackProperties
-
-    class Config:
-        extra = "forbid"  # No additional properties at top level
+    @field_validator('geometry')
+    @classmethod
+    def validate_track_geometry(cls, v):
+        """Ensure geometry is LineString or MultiLineString."""
+        if not isinstance(v, (LineString, MultiLineString)):
+            raise ValueError('Track features must have LineString or MultiLineString geometry')
+        return v
